updating course delete

+++ api/courses.js
@@ -1,7 +1,7 @@
 const { Router } = require("express");
 const { ValidationError } = require("sequelize");
 const { Course, CourseClientFields } = require("../models/course");
-const { User } = require("../models/user"); 
+const { User } = require("../models/user");
 const { requireAuthentication } = require("../lib/auth");
 const router = Router();
 
@@ -54,25 +54,24 @@ router.get("/", async (req, res, next) => {
 });
 
 //POST A COURSE
-router.post("/", requireAuthentication ,async (req, res, next) => {
+router.post("/", requireAuthentication, async (req, res, next) => {
   const admin = await req.admin;
-    if (admin) {
-      try {
-        const { subject, number, title, term, instructorID } = req.body;
-          if (!subject || !number || !title || !term || !instructorID) {
-          return res.status(400).send({ error: "Missing required fields" });
+  if (admin) {
+    try {
+      const { subject, number, title, term, instructorID } = req.body;
+      if (!subject || !number || !title || !term || !instructorID) {
+        return res.status(400).send({ error: "Missing required fields" });
       }
-        const course = await Course.create(req.body);
-        res.status(201).send({ id: course.id });
-      }
-      catch (e) {
-        next(e);
-      }
-    } else {
-      res.status(403).send({
-        error: "Unauthorized to create a course.",
-      });
+      const course = await Course.create(req.body);
+      res.status(201).send({ id: course.id });
+    } catch (e) {
+      next(e);
     }
+  } else {
+    res.status(403).send({
+      error: "Unauthorized to create a course.",
+    });
+  }
 });
 
 //GET A SPECIFIC COURSE
@@ -89,11 +88,11 @@ router.get("/:id", async (req, res, next) => {
 });
 
 //PATCH A SPECIFIC COURSE
-router.patch("/:id",requireAuthentication, async (req, res, next) => {
+router.patch("/:id", requireAuthentication, async (req, res, next) => {
   const instructorID = req.body.instructorID;
   const admin = await req.admin;
   console.log(admin);
-  if (req.user == instructorID || admin){
+  if (req.user == instructorID || admin) {
     try {
       const user = await User.findOne({
         where: { id: instructorID, role: "instructor" },
@@ -103,15 +102,14 @@ router.patch("/:id",requireAuthentication, async (req, res, next) => {
       }
       const courseId = req.params.id;
       const result = await Course.update(req.body, {
-          where: { id: courseId , instructorID: instructorID},
-          fields: CourseClientFields,
+        where: { id: courseId, instructorID: instructorID },
+        fields: CourseClientFields,
       });
       if (result[0] > 0) {
         res.status(200).send({
-          message: "Updated the course."
+          message: "Updated the course.",
         });
-      }
-      else{
+      } else {
         return res.status(404).send({
           error: "Invalid Instructor.",
         });
@@ -119,7 +117,7 @@ router.patch("/:id",requireAuthentication, async (req, res, next) => {
     } catch (e) {
       next(e);
     }
-  } else{
+  } else {
     return res.status(403).send({
       error: "Unauthorized To Access the specified resource",
     });
@@ -128,10 +126,17 @@ router.patch("/:id",requireAuthentication, async (req, res, next) => {
 
 // Only Authenticated user with admin can delete.
 router.delete("/:id", async (req, res, next) => {
-  const courseId = req.params.id;
-  const result = await Course.destroy({ where: { id: courseId } });
-  if (result > 0) {
-    res.status(204).send();
+  const admin = await req.admin;
+  if (admin) {
+    const courseId = req.params.id;
+    const result = await Course.destroy({ where: { id: courseId } });
+    if (result > 0) {
+      res.status(204).send();
+    } else {
+      res.status(404).send({ msg: "ID doesn't exist." });
+    }
+  } else {
+    res.status(403).send({ msg: "UnAuthorized access." });
   }
 });
